requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
filelock==3.16.1
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.22.0
networkx==3.4.2
numpy==2.2.1
openai==1.59.7
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pyarrow==19.0.0
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.36.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
pycryptodome==3.18.0  
